+++ lib/support/cancellation_exception.dart
@@ -1,0 +1,9 @@
+class CancellationException implements Exception {
+  @pragma("vm:entry-point")
+  const CancellationException([this.message = 'Operation is cancelled.']);
+
+  final String message;
+
+  @override
+  String toString() => "Exception: $message";
+}